# Speed up client-side bulk-handling