With this commit we speed up preparing bulk requests from data files by
implementing several optimizations:

* Bulk data are passed as a string instead of a list to the runner. This
avoids the cost of converting the raw list to a string in the Python
Elasticsearch client.
* Lines are read in bulk from the data source instead of line by line.
This avoids many method calls.
* We provide a special implementation for the common case (ids are
autogenerated, no conflicts) to make the hot code path as simple as
possible.

This commit also adds a microbenchmark that measures the speedup. The
following table shows a comparison of the throughput of the bulk reader
for various bulk sizes:

| Bulk Size | master [ops/s] | This PR [ops/s] | Speedup |
|-----------|----------------|-----------------|---------|
| 100       | 14829          | 92395           | 6.23    |
| 1000      | 1448           | 10953           | 7.56    |
| 10000     | 148            | 1100            | 7.43    |
| 100000    | 15             | 107             | 7.13    |

All data have been measured using Python 3.8 on Linux.